+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### v0.2.0
 
 - Support bandwidth limit on fftw.
+- Support limit the max traffic fftw can used every day.
 
 ### v0.1.0